requirements.txt
torch~=1.12.1
pyyaml~=6.0
tqdm~=4.64.1
opencv-python~=4.6.0.66
matplotlib~=3.6.2
numpy~=1.24.1
pandas~=1.5.2
seaborn~=0.12.1
pillow~=8.2.0
torchvision~=0.13.1
requests~=2.28.1
flask~=2.2.2
thop~=0.1.1-2209072238
einops~=0.6.0
gradio==4.44.1
fastapi==0.115.12
pydantic==2.10.6
uvicorn==0.33.0
mysql-connector-python==9.0.0